updated main.go and models directory
