+++ go.mod
@@ -1,0 +1,3 @@
+module github.com/tanggol69/linked-list/tree/doubly-linked-list
+
+go 1.20
